We don't need to expose tmpfs externally. Hide data structures!
